seperated tests and formatting
