+++ lib/cbus_elixir_web/templates/meeting/_paginate.html.eex
@@ -3,19 +3,19 @@
     <div>
     <%= if @paged_result.prev_page do %>
       <li>
-        <a href=<%= URI.merge(URI.parse("http://localhost:4000"), "meetings?page=#{1}") |> to_string() %>>
+        <a href=<%= URI.merge(URI.parse("https://www.columbuselixir.com/"), "meetings?page=#{1}") |> to_string() %>>
           <i class="fas fa-angle-double-left"></i>
         </a>
       </li>
       <li>
-        <a href=<%= URI.merge(URI.parse("http://localhost:4000"), "meetings?page=#{@paged_result.prev_page}") |> to_string() %>>
+        <a href=<%= URI.merge(URI.parse("https://www.columbuselixir.com/"), "meetings?page=#{@paged_result.prev_page}") |> to_string() %>>
           <i class="fas fa-angle-left"></i>
         </a>
       </li>
     <% end %>
 
      <li>
-        <a href=<%= URI.merge(URI.parse("http://localhost:4000"), "meetings?page=#{@paged_result.prev_page}") |> to_string() %>>
+        <a href=<%= URI.merge(URI.parse("https://www.columbuselixir.com/"), "meetings?page=#{@paged_result.prev_page}") |> to_string() %>>
           <h3><%= @paged_result.prev_page %></h3>
         </a>
       </li>
@@ -24,19 +24,19 @@
         <h3><%= @paged_result.page %></h3>
       </li>
       <li>
-        <a href=<%= URI.merge(URI.parse("http://localhost:4000"), "meetings?page=#{@paged_result.next_page}") |> to_string() %>>
+        <a href=<%= URI.merge(URI.parse("https://www.columbuselixir.com/"), "meetings?page=#{@paged_result.next_page}") |> to_string() %>>
           <h3><%= @paged_result.next_page %></h3>
         </a>
       </li>
     
     <%= if @paged_result.next_page do %>
       <li>
-        <a href=<%= URI.merge(URI.parse("http://localhost:4000"), "meetings?page=#{@paged_result.next_page}") |> to_string() %>>
+        <a href=<%= URI.merge(URI.parse("https://www.columbuselixir.com/"), "meetings?page=#{@paged_result.next_page}") |> to_string() %>>
           <i class="fas fa-angle-right"></i>
         </a>
       </li>
       <li>
-        <a href=<%= URI.merge(URI.parse("http://localhost:4000"), "meetings?page=#{@paged_result.last}") |> to_string() %>>
+        <a href=<%= URI.merge(URI.parse("https://www.columbuselixir.com/"), "meetings?page=#{@paged_result.last}") |> to_string() %>>
           <i class="fas fa-angle-double-right"></i>
         </a>
       </li>
