+++ test/test_helper.exs
@@ -1,3 +1,2 @@
-ExUnit.start()
-
 Ecto.Adapters.SQL.Sandbox.mode(CbusElixir.Repo, :manual)
+ExUnit.start()
